Add memory profiling to the go probe.
The go memory profile is in the same pprof format as the CPU profile, minus the compression.
Memory profiling is ALWAYS ON in a go program, capturing every 1/x memory allocation, but the default is for a low sample rate.

This commit captures the memory profile and merges the samples into the CPU profile data. It doesn't affect the counts, because they wouldn't make much sense. Instead, it uses the counts from the CPU profile, setting a minimum count of 1 to keep the BF parser happy.