Update the agent client to support sending blackfire.yml.
This requires altering how certain headers are structured, in what order they are sent, and switching between sending and receiving due to protocol design.

Tested end-to-end with a local installation and a dummy yml file.

The go probe only looks for `.blackfire.yml` in the same directory as the go app, and gives up sending the yml component if it's not found.